Add encoder input interface
